The new storage system is almost fully working now.  e.g. allNotes was not being cleared on opening a new file.  Also the velocity graph was not displaying correctly because of a float velocity being assigned to an int variable.  Also the customCompare for the sort in loadSequence was only comparing timeStamp and ignoring noteNumber.  Now sort is timeStamp as highest priority and then noteNumber as second priority.  The only remaining work is to enable humanization of time stamps and enable drag on note head to edit velocity and possibly time stamp.

+++ Source/NoteWithOffTime.h
@@ -45,7 +45,7 @@ public:
 //    triggers(-1),
 //    triggeredBy(-1),
 //    chainTrigger(-1),
-//    highestVelocityInChain(0),
+//    highestVelocityInChain(0.0f),
 //    triggeredNote(false),
 //    triggeredOffNote(false),
 //    autoplayedNote(false),
@@ -288,7 +288,7 @@ public:
     int triggers; //The step that this note triggers.  Set to -1 if last in group.
     int triggeredBy; //The that step that directly triggers this note.  Set to -1 if first in group.
     int chainTrigger; //The shortest note starting at the SAME time as the firstInChain. There may be other longer notes starting at the exact same time as the chainTrigger. The chainTrigger is not necessarily the firstInChain. Every step has a chainTrigger property including the chainTrigger itself.  A step numbered the same as its chainTrigger property is definitely a group trigger step.
-    int highestVelocityInChain; //Used for display of velocity graph.
+    float highestVelocityInChain; //Used for display of velocity graph.
     bool triggeredNote; //triggeredNotes are played when the chain trigger is played.  They are steps that start no later than the triggeredNoteLimit from the chainTrigger.  Unless they are triggeredOffNotes they are held at least their full duration but are extended if the chainTrigger is held beyond their scheduled end time, in which case they are held until the chainTrigger is released.
     bool triggeredOffNote; //triggeredOffNotes are triggeredNotes that are forced off when their chainTrigger turns off.  triggeredOffNotes are defined as notes that end before the END of the next chainTrigger note.
     bool autoplayedNote; //Notes that start more than triggeredNoteLimit after the chainTrigger.
